requirements.txt
addict==2.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
kiwisolver==1.4.7
Markdown==3.7
markdown-it-py==3.0.0
matplotlib==3.9.2
mdurl==0.1.2
mmcls==1.0.0rc3
mmengine==0.3.2
mmsegmentation==1.0.0rc1
model-index==0.1.11
modelindex==0.0.2
numpy==1.23.5
opencv-python-headless==4.7.0.68
openmim==0.3.3
ordered-set==4.1.0
packaging==24.1
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
prettytable==3.12.0
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
scipy==1.13.1
six==1.16.0
tabulate==0.9.0
termcolor==2.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
yapf==0.32.0
zipp==3.20.2
